@AutoWired 를 모두 @RequiredArgsConstructor변경 및 private final BoardService bService; 변수를 변경했음
